fixed package names vs physical directory.
